+++ CHANGELOG.rst
@@ -4,4 +4,4 @@ Changelog
 0.0.1 (2019-02-05)
 ------------------
 
-* Initial release of `django-exporter`
+* Initial release of `django-dataexporter`